Fix database stack creation
The test_slurm_accounting is testing the database stack template through a stack creation.
In this stack template two new subnets are being created and this creation cannot be done more than one time on a given VPC stack.
The fix is to create an ad-hoc VPC stack for this test

Signed-off-by: Luca Carrogu <carrogu@amazon.com>