Create test subnets in all the avail AZ
Create test subnets in all the avail AZ for a region
Also create child class of `CfnStack` called `CfnVpcStack`, which does have two methods to retrieve public and private subnets.

The public and private subnets returned by the `CfnVpcStack` are located in the default AZ, if the `DEFAULT_AVAILABILITY_ZONE` map contains an entry for the region, otherwise are in a random AZ.

The Jinja test variables `public_subnet_id` and `private_subnet_id` are set with the values retrieved from `get_public_subnet` and `get_private_subnet` exposed by the new class `CfnVpcStack`.

This change is an intermediate step to be able to specify an AZ as dimension of the test configuration.

Signed-off-by: Luca Carrogu <carrogu@amazon.com>