Fix logic for subnet with additional CIDR and no Internet
For backward compatibility, the subnet with additional CIDR must be created in an AZ different from the default subnet.
The same is valid for the subnet with no Internet.
In the future, it is expected that these two subnets will not be created anymore, since we are now creating a subnet for each AZ, but this would require to change the logic of the tests currently using them

Signed-off-by: Luca Carrogu <carrogu@amazon.com>