Improve wrong version error message in userdate
Old message: This AMI was created with aws-parallelcluster-cookbook-3.6.0, but is trying to be used with aws-parallelcluster-cookbook-3.5.0. Please either use an AMI created with aws-parallelcluster-cookbook-3.5.0 or change your ParallelCluster to aws-parallelcluster-cookbook-3.6.0

New message: This AMI was created with ParallelCluster 3.5.0, but is trying to be used with ParallelCluster 3.6.0. Please either use an AMI created with ParallelCluster 3.6.0 or change your ParallelCluster to 3.5.0

Signed-off-by: Hanwen <hanwenli@amazon.com>