Add rhel8 tests to daily integration tests
For some tests on which we are testing a single OS,
I replaced old OSes like centos7 and ubuntu18 favouring rhel8.

In future we need to enable not yet supported features like:
* test_ad_integration
* test_dcv
* test_cloudwatch_logging
* test_arm_pl

The following tests should be enabled but they are still failing:
* test_cluster_in_no_internet_subnet
* test_ebs_existing
* test_dynamic_file_systems_update

The following tests are not enabled because raid mount does not work yet
* test_raid_performance_mode
* test_update_slurm

The following tests are not enabled because system-analyzer fails
* test_disable_hyperthreading
* test_efa

Then in future we should remove the OSS_*_NO_RHEL8 temporary variables.
In future we can also replace centos7 with rhel8 in tests executed for a single OS.

Signed-off-by: Enrico Usai <usai@amazon.com>