[Integ-tests] Relax implicit timeout in test_slurm_scaling
It is a sporadic problem in integration tests code when testing in isolated regions (BTW, The sporadic failure has not happened in commercial regions).
There is an implicit timeout assuming static nodes can be replaced within 200 seconds. The speed of replacing compute nodes is not the purpose of the tests. Therefore, I extended the sleep job for 50 seconds. Therefore, the implicit timeout of static nodes replacement is increased by 50 seconds too.

Signed-off-by: Hanwen <hanwenli@amazon.com>