Improve loading time of nested templates
PyYAML builds a graph when loading files. This is time consumming, especially if there is no strong need to use the YAML format for the output.
In this case, the nested templates can be loaded as JSON using the standard python library). Since the main template is loaded as YAML, the complete CFN template on CLoudFormation Console will be rendered as YAML (the nested stack templates will be automatically converted to YAML when viewing the main stack).

This reduces loading time by an average of 5-6 seconds.

Signed-off-by: Eddy Mwiti <eddmwiti@amazon.com>