[CFNLimits] Use a snapshot to catch new sections/parameters in cluster configurations
As more cluster configuration sections are added, there may be more CFN Resources included in templates.
To ensure any that additions to the cluster configuration model are included in this test, this commit includes a snapshot of the full cluster config (with implied values).
The snapshot will be updated when new section/parameters are added.

Signed-off-by: Eddy Mwiti <eddmwiti@amazon.com>