Add redhat8 update cluster integration tests in develop
This patch adds rhel8 to the common cluster update integration tests in develop,
except for the awsbatch one, since awsbatch is not supported in rhel8, and the
multi-az one, since our ocr_stack class-scoped fixture has hardcoded
'Linux/UNIX' as platform, while rhel8 required 'Red Hat Enterprise Linux'.

If in a follow up patch we decide to set the platform value conditionally based
on the operating system, we ought to pay attention to what happens when running
this tests for, i.e., alinux2 and rhel8, since this fixture is shared (class
level scope) and the two OSs require different platform values.