Reenable validations for functionalities previously not supported in rhel8
Previously, intel mpi, arm pl and fabric manager were not supported in rhel8,
but we have added all these functionalities, therefore in this patch we reenable
the validation.

Note that in `parallelcluster_test.yaml` we skip some checks for scheduler
plugin in the case where the os is rhel8, because with rhel8 in the cookbook we
directly skip the installation of the scheduler plugin, therefore in the AMI we
don't have, i.e., the scheduler plugin python virtual environment.

Furthermore, we change the modulefiles regexp because in rhel8 it does not
appear in the `module avail` output.