Make test_slurm_memory_based_scheduling more robust
Reduce memory consumption of last job in the integration test to
minimize the risk for the job to run out of memory and be killed.

Set the new memory consumption of the job to 3 GB (3x10^9 B).

Change the Schedulable memory to 2500 MiB in the last phase of the
test to make sure the new `RealMemory` parameter is well below
the 3 GB used by the job.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>