Replace bash script with InSpec execution for ArmPL validation during AMI build
By reusing the same controls we use in Kitchen tests, we avoid code duplication and coupling between aws-parallelcluster and aws-parallelcluster-cookbooks.

This is the first step towards reusing InSpec controls for AMI validation and testing. We want all the validation and test checks to be replaced.

PatchInSpecProfiles is temporarily required because of a different cookbooks location in aws-parallelcluster-cookbooks and on the AMI.