Permit tags prefix validators to work if a tag is defined as a boolean
If one tag in the config has a boolean value (e.g. YES), tags variable will be a list of dict.
If all of them are strings, tags variable will be a list of BaseTag objects.

This patch improves the `validate_no_duplicate_tag` to be able to work in both the cases.

Previously, when using a config like:
```
Tags:
   - Key: WrongTagValue
     Value: YES
```
the cluster creation was failing with an error:
```
"message": "Invalid cluster configuration: 'dict' object has no attribute 'key'"
```

I replicated the issue within the unit tests and then fixed it.

Now, at cluster creation time, the validator is executed correctly and the Schema parsing fails with a meaningful message:
```
    "type": "ConfigSchemaValidator",
    "message": "[('Tags', {1: {'Value': ['Not a valid string.']}})]"
```

Signed-off-by: Enrico Usai <usai@amazon.com>