Do not check RAID File System at boot
We recently refactored `manage_raid` cookbook resource to use the common `volume` resource
and discovered that in past implementation we were setting `2` as File System Check Order.

0 means that fsck will not check the filesystem.
The root filesystem should be set to 1 and any others you want to be checked assigned after that.

We're changing the test logic to ensure the value is 0 rather than 2
because we do not want to check mounted file system at boot.
This can cause cluster timeout issues if the file system is too big (e.g. TiB)
and this is aligned to what we are doing with root volume, shared EBS, EFS and FSx.

### References

* https://www.redhat.com/sysadmin/etc-fstab

Signed-off-by: Enrico Usai <usai@amazon.com>