Propagate cluster s3 bucket information to the compute nodes
This is required to fetch config from S3 if for some reason
the config file is no longer available in the shared storage.

As part of cookbook refactoring I wrongly moved the fetch_config.rb
before the mount of the shared storage. Since the config was not available
from the shared storage, the compute node was trying to download it from s3
but the cluster/cluster_s3_bucket node attribute was unavailable.

I'm going to restore the order to avoid fetching from s3 every time
but in any case it's useful to add robustness to the compute nodes fetch_config logic.

Signed-off-by: Enrico Usai <usai@amazon.com>