Fix _test_update_queue_strategy_with_running_job
Use multi-node jobs in each queue instead of single-node, single-core
job arrays. This forces the `wait_job_running` method to wait for the
dynamic nodes to come up and start running the jobs.

Change assertion in the case of the TERMINATE queue parameter update
strategy to check the reason of the node instead of the state.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>