[Performance Test] Add scaling test to validate scale out to 1000 nodes.
The test is executed in 2 regions to collect more date.

With `-N 1000` we are sure to have 1000 nodes and using `srun` the job won't run
in the first available node available and it will wait for all the nodes.`

The test checks that clustermgtd logs does not contain bootstrap errors for compute nodes.

There are no static nodes to avoid a failure at cluster creation time.

Signed-off-by: Giacomo Marciani <mgiacomo@amazon.com>
Signed-off-by: Enrico Usai <usai@amazon.com>