Use pytest.raises in test_dcv to catch remote command errors when running on Ubuntu 22.04
It is unclear why running this test with Ubuntu 22.04 raises an exception.  Curl exits with a code of 56
when executing the test command for error cases, but returns 0 for the same on Ubuntu 20.  Potential differences
are openssl which is 3.0 and curl which is 7.81 on Ubuntu 22.  Tested updating the certificate generation to
use ed25519 for dcv but it made no difference.  Otherwise all positive tests for DCV are passing on Ubuntu 22.04
and able to launch clusters and open the web portal for DCV via dcv-connect.