Pass value of AssignPublicIp config param to LT associate_public_ip_address property
### Description of changes

We have the `MultiNetworkInterfacesInstancesValidator` that doesn't permit to set `AssignPublicIp: true`
when using a multi-NIC instance type, so this change does not have an impact in the default behaviour.

Anyway, this change permits to pass `AssignPublicIp: false` and this is useful
in cases where the user has a very strict PermissionsBoundary that explicitly forbids the creation
of EC2 instances unless you have in the request an explicit `associate_public_ip_address=false`.

### Tests

`AssignPublicIp: true`:
* standard instance type --> no changes
* multi-nic instance type --> blocked by `MultiNetworkInterfacesInstancesValidator` --> no changes

`AssignPublicIp: false`:
* standard instance type --> `associate_public_ip_address=False` (previously it was `None`) --> no changes in the behaviour
* multi-nic instance type --> `associate_public_ip_address=False` (previously it was `None`) --> no changes in the behaviour

Signed-off-by: Enrico Usai <usai@amazon.com>