Strip string configuration values
When using line folding AND there is an extra space at the end some parameters like
`PasswordSecretArn` are failing on both validation and cluster creation case.

### Tests

Before the patch (extra space at the end):
```
  PasswordSecretArn: >-
    arn:aws:secretsmanager:eu-west-1:xxx:secret:PasswordSecret-pcluster-ad-7WCRBu
```
it is failing with:
```
    {
      "level": "WARNING",
      "type": "PasswordSecretArnValidator",
      "message": "Cannot validate secret arn:aws:secretsmanager:eu-west-1:xxx:secret:PasswordSecret-pcluster-ad-7WCRBu . Please refer to ParallelCluster official documentation for more information."
    }
```
and cluster creation is failing with:
```
Expected process to exit with [0], but received '255'
---- Begin output of aws secretsmanager get-secret-value --secret-id arn:aws:secretsmanager:eu-west-1:xxx:secret:pcluster-ad-7WCRBu  --region eu-west --query 'SecretString' --output text ----
STDOUT:
STDERR: An error occurred (AccessDeniedException) when calling the GetSecretValue operation: User: arn:aws:sts::xxx:assumed-role/cluster-3-second-line-RoleHeadNode-yyy/i-xxx is not authorized to perform: secretsmanager:GetSecretValue on resource: arn:aws:secretsmanager:eu-west:xxx:secret:ReadOnly-xTbzmS because no identity-based policy allows the secretsmanager:GetSecretValue action
```

After the patch the validation is passing and creation succeeding.

### References
* https://yaml.org/spec/1.2.2/#65-line-folding

Signed-off-by: Enrico Usai <usai@amazon.com>