[Tests] Save managed storage ids and mark them for deletion once they are ready
In the update file system test we're setting `DeletionPolicy=Retain` for managed file system
then doing another update and the retained file systems should be removed
at test teardown by the `delete_storage_on_teardown` fixture.

If there is a crash during the last update, the test code might not arrive to the point
where it is marking these retained file systems to be deleted at teardown.

With this change we're moving the step where we're marking them as eligible for deletion at teardown
just after the update the is creating them.
This change should reduce the possibility to have leftover resources.

I also moved this code into a new `_retrieve_managed_storage_ids` function.

Signed-off-by: Enrico Usai <usai@amazon.com>