Modify describe_capacity_reservations to return a list of CapacityBlockReservationInfo
From this moment we don't need to know the structure of the describe-capacity-reservations
output in the entire code. We can use the `CapacityBlockReservationInfo` wrapper class
that decouples boto3 structure from parallelcluster internal logic.

The code is now more robust and thanks to this change I found a bug in the code:
the _validate_no_pg was returning no errors even if there were no capacity reservations
without PG.
Aligned the unit test for it.

Signed-off-by: Enrico Usai <usai@amazon.com>