Extend ComputeResourceSizeValidator to validate min count when using CB
At the moment all the instances in Capacity Block reservations are managed as static instances
so min count must be equal to max count and this value should be > 0.

To perform this check we had to move ComputeResourceSizeValidator registration at queue level
since capacity_type is specified at that level.

Signed-off-by: Enrico Usai <usai@amazon.com>