Add new CapacityReservationSizeValidator to verify number of instances is not exceeded
In the cluster config we created a map to keep track of the number of instances
associated through max-count to a given capacity reservation id.
Then we're verifying that the number of instances is not exceeding the TotalInstanceCount.

This will be executed for every capacity reservation, but has a strong value when using
Capacity Block reservation.

Signed-off-by: Enrico Usai <usai@amazon.com>