[Release 3.8] Add integ tests for a shared /home via FSx Lustre, FSx Ontap, FSx OpenZfs, and EBS
Updated the test_shared_home pytest function to check multiple filesystem types and create existing file systems as necessary
This is the initial effort to test across all FSes and FileCache will either be added to this list when
the issues with performance are resolved, or it will be excluded from mount options for /home if needed.

This test currently verifies mounting a new filesystem to a new cluster and will be improved in a follow up
PR to test mounting the same filesystem across multiple clusters.