Create integration tests for Capacity Blocks
The Capacity Blocks behaviour is totally simulated, with mocking and overrides.

Fleet-config and describe-capacity-reservations output will be mocked in the following way:
 - static nodes initially will be created because standard on-demand instances,
 - fleet-config.json will be modified to simulate there is a CB,
 - the describe_capacity_reservations will be mocked to return a pending/active CB to simulate state transition

The test will verify that Slurm reservations are created/deleted accordingly and job in pending will start.

This code relies on a new override mechanism introduced as part of https://github.com/aws/aws-parallelcluster-node/pull/602

Signed-off-by: Enrico Usai <usai@amazon.com>