[Develop] Add /home to internal EFS when not specified in SharedStorage (#5867)
We are going to include /home with the new `SharedStorageType` parameter to support a larger set of use cases.
With this change, when a user specifies `Efs` for `SharedStorageType` pcluster will replace all EBS exports with shared EFS storage.

FileCache is incompatible with the use case for shared /home.  The initial load times for files into the cache are too long
(over an hour in some cases) and cause timeouts and race conditions in the bootstrapping process.  For this reason, FileCache
is not allowed for mounting a shared /home.