Install kernel-devel for Rocky Linux from vault
kernel-devel should be aligned to the installed version.
If there is kernel 8.8 we should have kernel-devel 8.8.

Without this patch we were installing latest kernel-devel package.
We cannot use `--releasever` to install 8.8 version because it is not more available in the repo.
Previous releases are moved into a vault area once a new minor release version is available for at least a week.

On RHEL8 instead all versions of the packages will remain available.

With this patch we're downloading this specific package from vault and installing it
if it's not available in the repository for the specific release version.

Signed-off-by: Enrico Usai <usai@amazon.com>