Pin jsonschema for Lambda Layer to version 4.17.3
This was not necessary because we used Lambda runtime Python 3.9, which contains boto3-1.26.90, which does not rely on jsonschema. After using Lambda runtime Python 3.12, we are using a newer boto3 version and reply on jsonschema. We need to pin the jsonschema because of a known issue: https://github.com/aws/aws-cdk/issues/26300

Signed-off-by: Hanwen <hanwenli@amazon.com>