Allow traffic from external slurmdbd to slurmctld's
Add ingress rule in the external dbd client security group
to allow connections initiated from the slurmdbd.

Such connections may be established from the slurmdbd's side
when the slurmdbd drops and then recovers: in this case
it is the slurmdbd attempting to re-establish the connectivity
between itself and all the slurmctld's it was previously
connected to.

Remove useless egress rules from security groups (all outgoing
traffic is enabled by default at the moment).

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>