[Tests] Delete common.yaml integration tests config file
We have two different set of tests for released and develop version
but the differences are minimal because both develop.yaml and released.yaml are importing the same common.yaml.

With this commit we are just removing the common.yaml file
and duplicating the tests definition in develop.yaml and released.yaml file.
When specific dimensions are overridden in one of the two files I'm keeping the existing dimension.

In this way when adding a test for released version we’ll be forced to evaluate
if it’s really useful and properly define the dimensions for the different branches.

In a future commit will differentiate the dimensions set of tests and dimensions.

I also put the test-suites in alphabetical order.

Signed-off-by: Enrico Usai <usai@amazon.com>