[Tests] Remove test for scontrol update nodelist sorting bug
In Slurm 22.05.7 the `scontrol update node` logic was modified and a sorting routine was
introduced, which modified the order of the nodes in the nodelist,
this was causing mismatches between the Slurm nodenames and the assigned addresses.

The issue has been fixed in 22.05.8 release, included in Pcluster 3.5.0 so this test is no longer useful.

### References

https://github.com/aws/aws-parallelcluster/pull/4786
https://bugs.schedmd.com/show_bug.cgi?id=15731
Signed-off-by: Enrico Usai <usai@amazon.com>