[Tests] Remove test for dynamics nodes not powering down after an update
In Slurm 21.08 cloud nodes were not get powered-down after their
SuspendTime has expired if a cluster update is performed, which restarts the
slurmctld daemon and immediately performs a scontrol reconfigure.

This was causing a race condition between restart of slurmctld and scontrol reconfigure.

The issue has been fixed in Slurm 22.05.7 release, included in Pcluster 3.4.0, so this test is no longer useful.

### References

https://github.com/aws/aws-parallelcluster/pull/4746
https://bugs.schedmd.com/show_bug.cgi?id=13953

Signed-off-by: Enrico Usai <usai@amazon.com>