Fix PATH in external slurmdbd user data
In some OSes (e.g. Ubuntu) we have `cfn-init` available
via a python virtual environment managed by PCluster.
The path for cfn-init is somewhere in `/opt/parallelcluster`.

In other OSes (e.g. Amazon Linux 2) the `cfn-init` app
comes from an OS package and the path is under `/opt/aws`.

This makes sure the PATH is configured so that `cfn-init`
is found in all OSes.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>