[Tests] Reduce dimensions of DCV tests for released version
These tests are verifying user-data/cookbook code logic to configure/enable DCV
on instances with/without GPU, on different partitions and architectures.

Once released, code logic to configure and enable DCV is stable,
we don't need to test it on a daily basis.

We can keep a single test just as a canary for "DCV license endpoint" canary.

Signed-off-by: Enrico Usai <usai@amazon.com>