[Tests] Reduce EFA tests dimensions for released version
In develop we are testing all OSes families and different instance types:
c5n and p4d are used to verify performances, HPC instances and ARM architecture to verify setup.

Once released, code logic to configure EFA is stable, performances cannot change,
AMIs are stable, we don't need to test it on a daily basis.

We can keep p4d test as a canary for multi-nics instance AWS API behaviour.

Signed-off-by: Enrico Usai <usai@amazon.com>