[Tests] Remove Intel HPC test for released version
This test is verifying our logic on downloading and installing Intel HPC and Intel MPI at runtime,
if enabled by the user.

The packages are download from S3 and installed with `localinstall --cacheonly` options.
We have canaries for S3 artifacts, once released, code logic is stable, we don't need to test this on a daily basis.

Signed-off-by: Enrico Usai <usai@amazon.com>