[Tests] Remove Protected Mode test from released version
This is testing CF integration and code logic.
If compute nodes launch fails, the cluster will enter in Protected Mode,
and we want to validate that at creation time, this will cause a failure in cluster creation.

The behavior doesn't depend on the architecture,
OS, region, instance-type or any other dimension we may want to consider.

Once released, Slurm and node logic is stable, we don't need to test this on a daily basis.

Signed-off-by: Enrico Usai <usai@amazon.com>