[Tests] Remove scontrol reboot tests for released version
test_scontrol_reboot is focused on validating slurm's behavior for rebooting nodes.
Another part is checking that clustermgtd properly handles or ignores nodes that are rebooted.

test_scontrol_reboot_ec2_health_checks has been introduced to cover a bug in pcluster 3.4.1.
The test is OS and instance-type dependent, it is possible to reproduce the issue only with Ubuntu 20.04.
It verifies that scontrol reboot does not trigger node replacements due to EC2 health check failures.

Once released, Slurm and node logic is stable, we don't need to execute these tests on a daily basis.

### References
* https://github.com/aws/aws-parallelcluster/issues/4751

Signed-off-by: Enrico Usai <usai@amazon.com>