[Tests] Remove Slurm Accounting tests for released version
This test is checking that node deamons only manages Slurm partitions part of the cluster configuration as SlurmQueues.

This specifically verifies clustermgtd behavior when there is an extra partitions in the cluster
defined through `CustomSlurmSettings`, that should not be managed by ParallelCluster.

Once released, Slurm and code logic are stable, we don't need to execute this test on a daily basis.

Signed-off-by: Enrico Usai <usai@amazon.com>