[Tests] Reduce dimensions of FSx tests for released version
The checks performed in test_multiple_fsx are the same of test_fsx_lustre.

In develop is useful to test different dimensions to verify the setup and configuration of FSx is correct,
but once released, our AMIs and code logic are stable, we don't need to execute this test on a daily basis.

We can reduce dimensions and just keep one test as canary for AWS API behaviour.
test_fsx_lustre_configuration_options and test_fsx_lustre_backup are enough for this scope.

Signed-off-by: Enrico Usai <usai@amazon.com>