[Tests] Reduce dimensions of EFS tests for released version
The checks performed in test_efs_same_az is similar to test_multiple_efs.

In develop is useful to test different dimensions to verify the setup and configuration of EFS is correct,
but once released, our AMIs and code logic are stable, we don't need to execute this test on a daily basis.

We can reduce dimensions and just keep one test as canary for AWS API behaviour.

Signed-off-by: Enrico Usai <usai@amazon.com>