[Tests] Remove RAID tests for released version
The test verifies our logic and packages used to configure RAID are working as expected.

In develop is useful to test different dimensions to verify the setup and configuration is correct,
but once released, our AMIs and code logic are stable, we don't need to execute this test on a daily basis.

Signed-off-by: Enrico Usai <usai@amazon.com>