[Tests] Reduce dimensions of EBS tests for released version
The checks performed in test_ebs_single is similar to test_multiple_ebs.

In develop is useful to test different dimensions to verify the setup and configuration of EBS is correct,
but once released, our AMIs and code logic are stable, we don't need to execute this test on a daily basis.

We can reduce dimensions and just keep one test as canary for AWS API behaviour.

Signed-off-by: Enrico Usai <usai@amazon.com>