[Tests] Remove shared home test for released version
The test verifies our logic used to configure a shared home is working as expected
with different file system types: Lustre, Efs, OpenSxf, Ontap, etc.

We can continue to test all the dimensions on develop, but once released,
cookbook code is stable, we don't need to execute this test on a daily basis.

Signed-off-by: Enrico Usai <usai@amazon.com>