[Tests] Reduce dimensions of DNS tests
The test verifies dns/hostname settings when cluster DNS is disabled.

It was tested with all OSes, we can simplify and test it for different families
considering that DNS setup is also tested through Kitchen tests.

We don't need to test it for released version, because once released code logic is stable.

Signed-off-by: Enrico Usai <usai@amazon.com>