[Tests] Reduce dimensions for tests in no-internet subnets
This test creates a cluster in a subnet with no internet,
run simple integration test to check prolog and epilog
script failure, then run osu latency and checks that no failures occur.

We don't need to test it for released version, because once released, AMIs and CDK code are stable.
We can also avoid to test multiple times the same architecture for develop.

Signed-off-by: Enrico Usai <usai@amazon.com>