[Unit Test] Use temporary file for CLI logger test to avoid failures when executed in parallel
As GitHub actions we're executing multiple tests in parallel and all of them are writing in the same CLI log file.
This means we cannot assert the length of `log.readlines()` when using the default path for the CLI log.

Removed check when using default path and added a new test using PCLUSTER_LOG_FILE environment variable
to use a specific log file for each unit test.

This also adds coverage for the `PCLUSTER_LOG_FILE` feature.

Signed-off-by: Enrico Usai <usai@amazon.com>