[Tests] Improve createami test to validate and use the created AMI
By enabling DisableValidateAndTest we're going to execute all the Kitchen tests with `tag:install_` prefix
as part of the Image Builder validation phase.

We extended the test to use the generated AMI for a cluster and run a job on it.

We also extended the dimensions to run it on every supported OS and removed scheduler dimension.

Signed-off-by: Enrico Usai <usai@amazon.com>