[Integ Tests] Merge cfn-init tests and rename to test-essential-features
The idea is to have a single test to verify fundamental features.

In this case we have:
* Support for all the types of quotes in custom bootstrap scripts arguments
* Verify compute node is replaced on userdata failure

I had to modify the _test_replace_compute_on_failure code, to wait for a specific queue.
We cannot use soft assertion for this test because "wait_" functions are relying on assertion failures for retries.

Signed-off-by: Enrico Usai <usai@amazon.com>