Expose ExternalSlurmdbd as subschema with host and port
Expose ExternalSlurmdbd configuration as a subschema with two sub-
parameters, namely `Host` and `Port`, which map to `AccountingStorageHost`
and `AccountingStoragePort` in the `slurm.conf`.

Adapt unit tests and full example config to new schema.

Fix missing `DatabaseName` field in full example config.

Fix minor issue in a comment in the code.

Signed-off-by: Jacopo De Amicis <jdamicis@amazon.it>