[integ-tests] Remove tests for Rocky, Remove build-image tests for Ubuntu22
Rocky AMIs are not publicly released, therefore this commit removes tests for Rocky.
Ubuntu22 build-image test is failing because of Lustre client, therefore this commit change the test to use Ubuntu20.
This commit will only be applied to release-3.9 branch. We will improve the testing process on develop branch separately.

Signed-off-by: Hanwen <hanwenli@amazon.com>