Decouple region support from CLI release
Remove the constant variable and move the content to a file. This file will be uploaded to official S3 buckets during CLI release and/or region expansion.

If the file in S3 is not accessible, the code falls back to the local file. The file in S3 is not accessible when we are developing new versions or network access is limited

Signed-off-by: Hanwen <hanwenli@amazon.com>