[integ-tests] Fix test_create_wrong_pcluster_version
The failure was not signaled to CloudFormation via cfn-signal with version 3.6.1 because of this change https://github.com/aws/aws-parallelcluster/commit/aeb87fc9a7c453d18b470aa97abfa1e93b321979. In particular, for whatever version<3.10.0 the envar $CFN_BOOTSTRAP_VIRTUALENV_PATH is not set and so the cfn-signal command is not resolved.

Therefore, we change the version to 3.10.0 vs 3.11.0 to make ``$CFN_BOOTSTRAP_VIRTUALENV_PATH` setup consistent.

Signed-off-by: Hanwen <hanwenli@amazon.com>