[Integ-test] Reduce sporadic failures from `test_slurm_memory_based_scheduling`
Changed an error to warning because asserting job failures from insufficient memory is not the purpose of the test.
Rather, the test purpose is, when memory based scheduling is disabled, the total requested memory of jobs on a node can surpass the total memory; when memory based scheduling is enabled, the jobs wait until previous jobs finish.

When memory based scheduling is disabled, how operating systems handles two jobs requesting more memory than it has is not our concern.

Signed-off-by: Hanwen <hanwenli@amazon.com>