[integ-tests-framework] Set fallback for OS rotation when no supported AMI is available
Prior to this commit, for example, is alinux2 and alinux2023 are not available, the available OS for Batch would be 0. The test framework would fail for divided by 0 failure, and not run any tests.

This commit fall back to the supported OS for the feature. Although the tests will eventually fail because of lack of AMIs, the test framework wouldn't fail anymore

This commit has to do some code refactor to apply the change efficiently (in terms of lines of code)