[build-image] Move logic from validation to test phase
Validation phase is executed at the end of the image build on the same build instance. Test phase is executed after the image build and uses a separate test instance.

Prior to this commit, both phases could be switched on/off with the DevSetting parameter `DisableValidateAndTest`. After this commit, validation phase is no longer used. Test phase can be switched on/off with the same DevSetting parameter