Remove pinning of kernel-abi-stablelists
According to https://access.redhat.com/solutions/444773

With Red Hat Enterprise Linux 7 and 8, the stablelist is valid for the particular major release. This means that once a symbol has been introduced into kABI for a particular major release, it will not be removed, nor will its meaning be changed during that kernel major release complete life cycle.

With Red Hat Enterprise Linux 9, each minor release will have a unique stablelist that is valid throughout the minor release lifecycle. For more information on this, please refer to the following knowledgebase article;

We are pinning  the minor release for the duration of the build. So the stablelist for the minor version will be installed