Echo chef-client.log when a node fails to bootstrap
This helps with investigating bootstrap failures in cases CloudWatch logs are not available.

Older CINC was outputting logs to STDOUT even when --logfile is specified. i.e. CINC was outputting logs to both STDOUT and the log file.

Sometime in CINC 16, it stopped outputting logs to STDOUT when --logfile is specified. There are no clear changelog items. But there are some discussions around it:
https://discourse.chef.io/t/chef-log-not-printing-for-console-in-chef-client-16-6-14/17983