Revert "[Test] Execute test_build_image on both x86 and arm instance types, both for develop and released branch."
This reverts commit fe6a186e9957ba7e85f83c1774225d83326c93a4.

This is a temporary revert to attempt to fix intermittent build image failure with "InvalidInstanceId". We will observe the test for a few days. If the revert fixes the issue, it means it is an API throttling issue, and we will raise the API throttling limit and reintroduce the commit

Reverting the commit is the easiest way to test because this suspected throttling only happens when the whole suite of develop.yaml is run. i.e. the intermittent "InvalidInstanceId" failure doesn't happen when running only the build image tests