[integ-tests] Add baseline for new OSes in test_startup_time
Before this commit, the tests on new OSes were failing with `KeyError`. This commits add baseline to the new OSes. The numbers of alinux2023 is copied from alinux2, rhel9 is copied from rhel8, rocky9 is copied from rocky8, ubuntu24 is copied from ubuntu22. We can change the number more accurately in the future. This commit aims at not failing the test with `KeyError`.