[integ-tests] With test_gb200, only run MPI check when instance is GB200
The MPI check on g5g takes 35 seconds to run. The current timeout is 20 seconds, causing the MPI job to fail. The purpose of test_gb200 is to test GB200 working properly. Testing everything about g5g is not the purpose of this test. Therefore, instead of investigating on the slowness of MPI on g5g, this commit only runs MPI when the instance is GB200.

MPI jobs are tested in other tests with other instance types