[integ-tests] Let serial_execution_by_instance to take effect on all p-series instances
Fixture serial_execution_by_instance can be optional used with each test. If the fixture is not used, the test could still run p-series instances in parallel.

This commit fixes a "bug" that some p-series instances serial execution is not enforced even when serial_execution_by_instance is used in the test