[integ-tests-framework] Add compatibility with local zones
Prior to this commit, integration tests framework wouldn't set up any network infrastructure in local zones.

This commit creates a public subnet, a private subnet in local zone. And route network from the private subnet to a NAT gateway in the parent region. NAT gateway is not available in local zones